requirements.txt
Flask
pandas
openpyxl
opencv-python
face_recognition
qrcode
Pillow
numpy
gunicorn